# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Inventory Management System REST API. Every response is wrapped in a `data` envelope; errors use an `error` envelope with a stable machine-readable code.",
        "title": "IMS API",
        "contact": {},
        "version": "1.0.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/purchase-orders": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "purchase-orders"
                ],
                "summary": "List purchase orders",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by status",
                        "name": "filter[status]",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_sabin-bhattarai_ims-backend_internal_shared.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/internal_purchasing.PurchaseOrder"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/definitions/github_com_sabin-bhattarai_ims-backend_internal_shared.PageMeta"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "purchase-orders"
                ],
                "summary": "Create a purchase order",
                "parameters": [
                    {
                        "description": "Purchase order",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_purchasing.POInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_sabin-bhattarai_ims-backend_internal_shared.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_purchasing.PurchaseOrder"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/purchase-orders/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "purchase-orders"
                ],
                "summary": "Get a purchase order",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Purchase order id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_sabin-bhattarai_ims-backend_internal_shared.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_purchasing.PurchaseOrder"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/purchase-orders/{id}/approve": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "purchase-orders"
                ],
                "summary": "Approve a purchase order",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Purchase order id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_sabin-bhattarai_ims-backend_internal_shared.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_purchasing.PurchaseOrder"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_sabin-bhattarai_ims-backend_internal_shared.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/purchase-orders/{id}/cancel": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "purchase-orders"
                ],
                "summary": "Cancel a purchase order",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Purchase order id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_sabin-bhattarai_ims-backend_internal_shared.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_purchasing.PurchaseOrder"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/purchase-orders/{id}/receipts": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "purchase-orders"
                ],
                "summary": "List goods receipts for a purchase order",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Purchase order id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_sabin-bhattarai_ims-backend_internal_shared.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/internal_purchasing.GoodsReceipt"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/purchase-orders/{id}/receive": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Supports partial receipt; the order becomes partially_received until every line is satisfied.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "purchase-orders"
                ],
                "summary": "Receive goods against a purchase order",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Purchase order id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Received lines",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_purchasing.ReceiveInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_sabin-bhattarai_ims-backend_internal_shared.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_purchasing.GoodsReceipt"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/github_com_sabin-bhattarai_ims-backend_internal_shared.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/purchase-orders/{id}/reject": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "purchase-orders"
                ],
                "summary": "Reject a purchase order",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Purchase order id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Rejection reason",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_purchasing.RejectInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_sabin-bhattarai_ims-backend_internal_shared.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_purchasing.PurchaseOrder"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/purchase-orders/{id}/submit": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "purchase-orders"
                ],
                "summary": "Submit a purchase order for approval",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Purchase order id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_sabin-bhattarai_ims-backend_internal_shared.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_purchasing.PurchaseOrder"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/github_com_sabin-bhattarai_ims-backend_internal_shared.ErrorEnvelope"
                        }
                    }
                }
            }
        },
...
        "/suppliers": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "suppliers"
                ],
                "summary": "List suppliers",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_sabin-bhattarai_ims-backend_internal_shared.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/internal_purchasing.Supplier"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/definitions/github_com_sabin-bhattarai_ims-backend_internal_shared.PageMeta"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "suppliers"
                ],
                "summary": "Create a supplier",
                "parameters": [
                    {
                        "description": "Supplier",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_purchasing.SupplierInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_sabin-bhattarai_ims-backend_internal_shared.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_purchasing.Supplier"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_sabin-bhattarai_ims-backend_internal_shared.Envelope": {
            "type": "object",
            "properties": {
                "data": {},
                "meta": {}
            }
        },
...
        "github_com_sabin-bhattarai_ims-backend_internal_shared.ErrorEnvelope": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/github_com_sabin-bhattarai_ims-backend_internal_shared.Error"
                }
            }
        },
        "github_com_sabin-bhattarai_ims-backend_internal_shared.PageMeta": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "per_page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
        "internal_purchasing.POInput": {
            "type": "object",
            "required": [
                "lines",
                "supplier_id",
                "warehouse_id"
            ],
            "properties": {
                "currency": {
                    "type": "string"
                },
                "expected_date": {
                    "type": "string"
                },
                "lines": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/internal_purchasing.POLineInput"
                    }
                },
                "note": {
                    "type": "string",
                    "maxLength": 2000
                },
                "shipping_total": {
                    "type": "number"
                },
                "supplier_id": {
                    "type": "string"
                },
                "warehouse_id": {
                    "type": "string"
                }
            }
        },
...
        "internal_purchasing.PurchaseOrder": {
            "type": "object",
            "properties": {
                "approved_at": {
                    "type": "string"
                },
                "approved_by": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "discount_total": {
                    "type": "number"
                },
                "expected_date": {
                    "type": "string"
                },
                "grand_total": {
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "lines": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_purchasing.PurchaseOrderLine"
                    }
                },
                "note": {
                    "type": "string"
                },
                "organization_id": {
                    "type": "string"
                },
                "rejected_at": {
                    "type": "string"
                },
                "rejected_by": {
                    "type": "string"
                },
                "rejection_reason": {
                    "type": "string"
                },
                "shipping_total": {
                    "type": "number"
                },
                "status": {
                    "$ref": "#/definitions/internal_purchasing.Status"
                },
                "submitted_at": {
                    "type": "string"
                },
                "subtotal": {
                    "type": "number"
                },
                "supplier": {
                    "$ref": "#/definitions/internal_purchasing.Supplier"
                },
                "supplier_id": {
                    "type": "string"
                },
                "tax_total": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                },
                "warehouse_id": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Access token as \"Bearer \u003ctoken\u003e\". Obtain one from /auth/login and refresh it at /auth/refresh.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}